requirements.txt
annotated-types==0.6.0
anyio==3.7.1
astroid==3.0.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
click==8.1.7
cryptography==41.0.4
dill==0.3.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.103.2
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
orjson==3.9.7
passlib==1.7.4
platformdirs==3.11.0
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyMySQL==1.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomlkit==0.12.1
typing_extensions==4.8.0
ujson==5.8.0
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
